requirements.txt
pikepdf
msoffcrypto-tool
paramiko
tqdm
cryptography
bcrypt
pynacl
cffi
Pillow==10.2.0
python-magic==0.4.27
exifread==3.0.0
stegano==0.11.1
rich==13.7.0
alive-progress==3.1.5